Jeremiah 29:15,20: Fix reordering of "in Babylon"
McFadyen had relocated v15 to after v20 (see p108 of book, and appendix
p208).
(The appendix doesn't seem to explain why this verse specifically was
moved, but does say that v14 and v16-20 probably weren't in the
original text.)

The first line of v15 had been moved back to its rightful place,
but the remainder of the verse ("in Babylon.") had not been moved.