Jeremiah 39: fix verse reordering.
McFadyen had moved 39:3,14 to the start of the chapter, along with a copy of
part of the preceding verse, 38:28. (See p140, appendix page p212.  It
says that v4-13 aren't in the Septuagint.)

The whole lines of the verses had been moved back, but the tails remained
where they were.

38:28b only contains duplication of the second part of 28a.
("28a" actually contains the whole of verse 28, and "28b" effectively
duplicates part of it.  "all the chief officers of" wasn't part of v28,
and belongs in 39:3.)

Move the section heading ("Provision made for Jeremiah's safety ...") to
just before v14, even though it creates a one-verse section.