More Jeremiah fixups (#328)
* Jeremiah: Fix more typos.  As before, these are generally not present in
the source material (see
https://archive.org/details/jeremiahinmodern00mcfauoft ),
and look like keyboard slips, presumably from manual transcription.

* Jeremiah 52:2 - Add missing line

This line is present in the source material (McFadyen p187), so I'm
guessing the transcriber just missed it.  Perhaps because it starts with
the same word as the following line.

* Jeremiah: Reposition verse markers in prose.

(McFadyen always placed verse markers at the start of the line, presumably for readability,
but where the verse does not begin a new paragraph, the division usually appears somewhere
within the line.)

* Jeremiah 15:8-9: Recombine verses (see McFadyen p67-68)

McFadyen relocated 8a/b ("Their widows are more in number / than the sand
of all the seas") to before 9e/f.
I'm not sure why, and the appendix (p201) didn't really seem to help.

* Jeremiah: add linebreaks in text where McFadyen has done.

* Jeremiah: fix instances of prose mistaken for poetry

* Jeremiah: fix instances of poetry mistakenly put as prose

* Jeremiah 24:3 (p95), 44:11 (p152): remove a paragraph break, add a paragraph break

* Jeremiah: Some verse indentation fixups.
Not intended as an exhaustive list, just ones I spotted while
comparing against the source text
(https://archive.org/details/jeremiahinmodern00mcfauoft )
3:25 (p25)
5:25 (p34)
6:14 (p37)
31:33 (p118)

* Jeremiah 10: move title from start of chapter to verse 17.

Verses 17-25 had been relocated by McFadyen to the start of chapter 10.
The chunk has been moved back to the end of the chapter, but the title had
not been moved with it.

* Jeremiah 29:15,20: Fix reordering of "in Babylon"

McFadyen had relocated v15 to after v20 (see p108 of book, and appendix
p208).
(The appendix doesn't seem to explain why this verse specifically was
moved, but does say that v14 and v16-20 probably weren't in the
original text.)

The first line of v15 had been moved back to its rightful place,
but the remainder of the verse ("in Babylon.") had not been moved.

* Jeremiah 39: fix verse reordering.

McFadyen had moved 39:3,14 to the start of the chapter, along with a copy of
part of the preceding verse, 38:28. (See p140, appendix page p212.  It
says that v4-13 aren't in the Septuagint.)

The whole lines of the verses had been moved back, but the tails remained
where they were.

38:28b only contains duplication of the second part of 28a.
("28a" actually contains the whole of verse 28, and "28b" effectively
duplicates part of it.  "all the chief officers of" wasn't part of v28,
and belongs in 39:3.)

Move the section heading ("Provision made for Jeremiah's safety ...") to
just before v14, even though it creates a one-verse section.

* Jeremiah 33:11, 38:22 - centrally align some poetry/quotes
(as in McFadyen p125,140)

* Jeremiah: centrally align some text (as McFadyen has done):

23:9 "Concerning the prophets." (McFadyen, p91)
46:2 "On Egypt." (p156)
49:7 "On Edom." (p167)
51:64 "Here end the words of Jeremiah." (p187)